v4/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.108.0
uvicorn>=0.25.0
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.0
streamlit>=1.28.0